@@ -1,9 +1,19 @@
-module tcw.im/tdi
+module pkg.tcw.im/tdi
 
-go 1.15
+go 1.20
 
 require (
-	github.com/labstack/echo/v4 v4.9.0
-	tcw.im/go-disk-usage v0.1.1
-	tcw.im/gtc v1.0.0
+	github.com/labstack/echo/v4 v4.11.4 // indirect
+	github.com/labstack/gommon v0.4.2 // indirect
+	github.com/mattn/go-colorable v0.1.13 // indirect
+	github.com/mattn/go-isatty v0.0.20 // indirect
+	github.com/shirou/gopsutil v3.21.11+incompatible // indirect
+	github.com/valyala/bytebufferpool v1.0.0 // indirect
+	github.com/valyala/fasttemplate v1.2.2 // indirect
+	golang.org/x/crypto v0.17.0 // indirect
+	golang.org/x/net v0.19.0 // indirect
+	golang.org/x/sys v0.15.0 // indirect
+	golang.org/x/text v0.14.0 // indirect
+	pkg.tcw.im/go-disk-usage v1.0.0 // indirect
+	pkg.tcw.im/gtc v1.1.0 // indirect
 )
